Android/SDL buildfix, move remains of the old UI into other files.
