+++ CMakeLists.txt
@@ -1130,7 +1130,6 @@ set(NativeAppSource
 	UI/GameInfoCache.cpp
 	UI/MainScreen.cpp
 	UI/MiscScreens.cpp
-	UI/MenuScreens.cpp
 	UI/GameScreen.cpp
 	UI/GameSettingsScreen.cpp
 	UI/GamepadEmu.cpp
