+++ UI/EmuScreen.cpp
@@ -46,7 +46,6 @@
 #include "UI/UIShader.h"
 
 #include "UI/MainScreen.h"
-#include "UI/MenuScreens.h"
 #include "UI/EmuScreen.h"
 #include "UI/GameInfoCache.h"
 #include "UI/MiscScreens.h"
