+++ UI/MiscScreens.cpp
@@ -27,7 +27,6 @@
 #include "ui/view.h"
 #include "ui/viewgroup.h"
 #include "UI/MiscScreens.h"
-#include "UI/MenuScreens.h"
 #include "UI/EmuScreen.h"
 #include "UI/MainScreen.h"
 #include "Core/Config.h"
@@ -41,7 +40,93 @@
 #pragma execution_character_set("utf-8")
 #endif
 
-void DrawBackground(float alpha);
+
+#include "base/timeutil.h"
+#include "base/colorutil.h"
+#include "gfx_es2/draw_buffer.h"
+#include "gfx_es2/gl_state.h"
+#include "util/random/rng.h"
+
+#include "Core/HLE/sceUtility.h"
+#include "UI/ui_atlas.h"
+#include "UI/ui.h"
+
+static const int symbols[4] = {
+	I_CROSS,
+	I_CIRCLE,
+	I_SQUARE,
+	I_TRIANGLE
+};
+
+static const uint32_t colors[4] = {
+	0xC0FFFFFF,
+	0xC0FFFFFF,
+	0xC0FFFFFF,
+	0xC0FFFFFF,
+};
+
+void DrawBackground(float alpha) {
+	static float xbase[100] = {0};
+	static float ybase[100] = {0};
+	static int last_dp_xres = 0;
+	static int last_dp_yres = 0;
+	if (xbase[0] == 0.0f || last_dp_xres != dp_xres || last_dp_yres != dp_yres) {
+		GMRng rng;
+		for (int i = 0; i < 100; i++) {
+			xbase[i] = rng.F() * dp_xres;
+			ybase[i] = rng.F() * dp_yres;
+		}
+		last_dp_xres = dp_xres;
+		last_dp_yres = dp_yres;
+	}
+	glstate.depthWrite.set(GL_TRUE);
+	glstate.colorMask.set(GL_TRUE, GL_TRUE, GL_TRUE, GL_TRUE);
+	glClearColor(0.1f,0.2f,0.43f,1.0f);
+	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT);
+	ui_draw2d.DrawImageStretch(I_BG, 0, 0, dp_xres, dp_yres);
+	float t = time_now();
+	for (int i = 0; i < 100; i++) {
+		float x = xbase[i];
+		float y = ybase[i] + 40*cos(i * 7.2 + t * 1.3);
+		float angle = sin(i + t);
+		int n = i & 3;
+		ui_draw2d.DrawImageRotated(symbols[n], x, y, 1.0f, angle, colorAlpha(colors[n], alpha * 0.1f));
+	}
+}
+
+std::map<std::string, std::pair<std::string, int>> GetLangValuesMapping() {
+	std::map<std::string, std::pair<std::string, int>> langValuesMapping;
+	langValuesMapping["ja_JP"] = std::make_pair("日本語", PSP_SYSTEMPARAM_LANGUAGE_JAPANESE);
+	langValuesMapping["en_US"] = std::make_pair("English",PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["fr_FR"] = std::make_pair("Français", PSP_SYSTEMPARAM_LANGUAGE_FRENCH);
+	langValuesMapping["es_ES"] = std::make_pair("Castellano", PSP_SYSTEMPARAM_LANGUAGE_SPANISH);
+	langValuesMapping["es_LA"] = std::make_pair("Latino", PSP_SYSTEMPARAM_LANGUAGE_SPANISH);
+	langValuesMapping["de_DE"] = std::make_pair("Deutsch", PSP_SYSTEMPARAM_LANGUAGE_GERMAN);
+	langValuesMapping["it_IT"] = std::make_pair("Italiano", PSP_SYSTEMPARAM_LANGUAGE_ITALIAN); 
+	langValuesMapping["nl_NL"] = std::make_pair("Nederlands", PSP_SYSTEMPARAM_LANGUAGE_DUTCH);
+	langValuesMapping["pt_PT"] = std::make_pair("Português", PSP_SYSTEMPARAM_LANGUAGE_PORTUGUESE);
+	langValuesMapping["pt_BR"] = std::make_pair("Brasileiro", PSP_SYSTEMPARAM_LANGUAGE_PORTUGUESE);
+	langValuesMapping["ru_RU"] = std::make_pair("Русский", PSP_SYSTEMPARAM_LANGUAGE_RUSSIAN);
+	langValuesMapping["ko_KR"] = std::make_pair("한국어", PSP_SYSTEMPARAM_LANGUAGE_KOREAN);
+	langValuesMapping["zh_TW"] = std::make_pair("繁體中文", PSP_SYSTEMPARAM_LANGUAGE_CHINESE_TRADITIONAL);
+	langValuesMapping["zh_CN"] = std::make_pair("简体中文", PSP_SYSTEMPARAM_LANGUAGE_CHINESE_SIMPLIFIED);
+
+	//langValuesMapping["ar_AE"] = std::make_pair("العربية", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["az_AZ"] = std::make_pair("Azeri", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["ca_ES"] = std::make_pair("Català", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["gr_EL"] = std::make_pair("ελληνικά", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["he_IL"] = std::make_pair("עברית", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["hu_HU"] = std::make_pair("Magyar", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["id_ID"] = std::make_pair("Indonesia", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["pl_PL"] = std::make_pair("Polski", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["ro_RO"] = std::make_pair("Român", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["sv_SE"] = std::make_pair("Svenska", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["tr_TR"] = std::make_pair("Türk", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	langValuesMapping["uk_UA"] = std::make_pair("Українська", PSP_SYSTEMPARAM_LANGUAGE_ENGLISH);
+	return langValuesMapping;
+}
+
+
 
 void UIScreenWithBackground::DrawBackground(UIContext &dc) {
 	::DrawBackground(1.0f);
@@ -384,6 +469,7 @@ void CreditsScreen::render() {
 		}
 		y += itemHeight;
 	}
+
 	/*
 	I18NCategory *g = GetI18NCategory("General");
