+++ UI/UI.vcxproj
@@ -26,7 +26,6 @@
     <ClCompile Include="GameScreen.cpp" />
     <ClCompile Include="GameSettingsScreen.cpp" />
     <ClCompile Include="MainScreen.cpp" />
-    <ClCompile Include="MenuScreens.cpp" />
     <ClCompile Include="MiscScreens.cpp" />
     <ClCompile Include="NativeApp.cpp" />
     <ClCompile Include="OnScreenDisplay.cpp" />
@@ -43,7 +42,6 @@
     <ClInclude Include="GameSettingsScreen.h" />
     <ClInclude Include="Header.h" />
     <ClInclude Include="MainScreen.h" />
-    <ClInclude Include="MenuScreens.h" />
     <ClInclude Include="MiscScreens.h" />
     <ClInclude Include="OnScreenDisplay.h" />
     <ClInclude Include="PluginScreen.h" />
